Added non-vanilla function blam_collision_surface_test3d to collision.c
This function performs a signed volume test to determine if a ray intersects a
convex polygon.

The intent was to see if the projection step of blam_collision_surface_test2d was
causing discrepancies when used as mitigation for phantom BSP.

Result: blam_collision_surface_test3d adds no meaningful information. Back to the
drawing board.