note about next step for phantom BSP resolution
This next phase involves speculation on BSP using blam_collision_surface_test2d. If
vanilla Halo would return a surface hit but the surface test fails, then the surface
might be phantom BSP.

If the surface might be phantom BSP, the BSP is traversed out to infinity, as if the
surface was not hit. If the next surface hit (or no next surface is hit) has a
facing that is consistent with the speculated surface, then the speculated surface
is accepted as an actual surface.

There are two approaches to implementing this:
    1. continue the current BSP traversal; or,
    2. duplicate and specialize the BSP traversal subroutines.

The first approach is more efficient but much more intrusive. The second approach
is not quite ideal in terms of efficiency but is easier to understand and insert.

I will implement the first approach at some later point, if I have the time.