phantom BSP: even better technique to eliminate phantom BSP
This technique is based off of the previous technique, which performs a quick
validation test followed by a test which observes if the next surface orientation
is consistent with valid surface, should the quick test fail.

However, rather than test the next surface orientation, observe that if one
encounters front-facing phantom BSP, then the next solid partition must be a BSP
leak. Coincidentally, the failures of the previous technique are the same failures
of this technique; if the next solid partition is a proper leak where there
is valid surface data, then intersections that fail the quick validation test will
incorrectly classify the speculated surface as phantom BSP.

This implementation is not yet complete; back-facing phantom BSP is not handled
correctly yet. If we encounter back-facing phantom BSP, then the previous solid
partition crossed must be a BSP leak.

Additionally, the secondary validation test may be enhanced by testing the absolute
distance between the candidate surface and the leak. If this distance exceeds some
threshold, chances are that the leak is a proper leak where surface data should be.
This enhancement should be used as a stopgap until BSP leaks are properly solved.

Furthermore, mitigation technique is more integrated into the BSP-vector
intersection test. As long as there is a pending surface result, the BSP-vector
interseciton test should continue out to infinity if need be. The chances of this
mattering is small, so this isn't high on the priority list.